backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow
numpy
pillow
starlette